Change DB sum type to numeric for better cross-locale compability

+++ src/server/data/Categories.ts
@@ -64,8 +64,8 @@ function getTotals(tx: IBaseProtocol<any>) {
       `
 SELECT
   categories.id, categories.parent_id as "parentId",
-  SUM(CASE WHEN type='expense' AND template=false AND date >= $/startDate/::DATE AND date < $/endDate/::DATE THEN sum::NUMERIC ELSE 0::NUMERIC END) AS expenses,
-  SUM(CASE WHEN type='income' AND template=false AND date >= $/startDate/::DATE AND date < $/endDate/::DATE THEN sum::NUMERIC ELSE 0::NUMERIC END) AS income
+  SUM(CASE WHEN type='expense' AND template=false AND date >= $/startDate/::DATE AND date < $/endDate/::DATE THEN sum ELSE 0::NUMERIC END) AS expenses,
+  SUM(CASE WHEN type='income' AND template=false AND date >= $/startDate/::DATE AND date < $/endDate/::DATE THEN sum ELSE 0::NUMERIC END) AS income
 FROM categories
 LEFT JOIN expenses ON categories.id = category_id
 WHERE categories.group_id = $/groupId/
